updated abby quote order

+++ abby_quote_order/clean_all.sh
@@ -1,14 +1,47 @@
 #!/bin/bash
+#==============================================================================
+#
+#	THIS FILE CLEANS ALL VARIABLES AND FILES
+#
+#==============================================================================
 
-if [ -z $CONFIG ]
-then
-	source config.sh
+#	SOURCE CONFIG FILE
+if [ -z $CONFIG ]; then source config.sh ; fi
+
+#	REMOVE EXTRA FILES
+if [[ $1 ]]; then
+	echo "clean_all.sh 1"
+	rm $JSONPROBS
+	rm $JSONKEYWORDARRAYS
+	rm $PROBSFILE
+	rm $SORTEDPROBS
+	rm $SORTEDID
+else
+	echo "clean_all.sh"
 fi
 
-rm $JSONPROBS
-rm $JSONKEYWORDARRAYS
-rm $PROBSFILE
-rm $SORTEDPROBS
-rm $SORTEDID
-
+#	UNSET ALL VARIABLES
+unset ABBYQUOTEORDER
+unset CNFIG
+unset https
+unset ssID
+unset altjson
+unset SHEETID
+unset PROBSENTRY
+unset JSON
+unset PROBS
+unset JSONKEYWORDARRAYS
+unset PROBSFILE
+unset ARRPREFIX
+unset ARRDIR
+unset COUNT
+unset DATE
+unset SORTEDPROBS
+unset SORTEDID
 unset CONFIG
+
+#==============================================================================
+#
+#	END CLEAN ALL
+#
+#==============================================================================
